Struct rusoto_cognito_idp::CreateUserPoolClientRequest

pub struct CreateUserPoolClientRequest {
    pub allowed_o_auth_flows: Option<Vec<String>>,
    pub allowed_o_auth_flows_user_pool_client: Option<bool>,
    pub allowed_o_auth_scopes: Option<Vec<String>>,
    pub callback_ur_ls: Option<Vec<String>>,
    pub client_name: String,
    pub default_redirect_uri: Option<String>,
    pub explicit_auth_flows: Option<Vec<String>>,
    pub generate_secret: Option<bool>,
    pub logout_ur_ls: Option<Vec<String>>,
    pub read_attributes: Option<Vec<String>>,
    pub refresh_token_validity: Option<i64>,
    pub supported_identity_providers: Option<Vec<String>>,
    pub user_pool_id: String,
    pub write_attributes: Option<Vec<String>>,
}

Represents the request to create a user pool client.

Fields
allowed_o_auth_flows: Option<Vec<String>>

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint.

Set to token to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

allowed_o_auth_flows_user_pool_client: Option<bool>

Set to True if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

allowed_o_auth_scopes: Option<Vec<String>>

A list of allowed OAuth scopes. Currently supported values are "phone", "email", "openid", and "Cognito".

callback_ur_ls: Option<Vec<String>>

A list of allowed callback URLs for the identity providers.

client_name: String

The client name for the user pool client you would like to create.

default_redirect_uri: Option<String>

The default redirect URI. Must be in the CallbackURLs list.

explicit_auth_flows: Option<Vec<String>>

The explicit authentication flows.

generate_secret: Option<bool>

Boolean to specify whether you want to generate a secret for the user pool client being created.

logout_ur_ls: Option<Vec<String>>

A list of allowed logout URLs for the identity providers.

read_attributes: Option<Vec<String>>

The read attributes.

refresh_token_validity: Option<i64>

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

supported_identity_providers: Option<Vec<String>>

A list of provider names for the identity providers that are supported on this client.

user_pool_id: String

The user pool ID for the user pool where you want to create a user pool client.

write_attributes: Option<Vec<String>>

The write attributes.