Struct rusoto_cognito_idp::CreateUserPoolRequest

pub struct CreateUserPoolRequest {
    pub admin_create_user_config: Option<AdminCreateUserConfigType>,
    pub alias_attributes: Option<Vec<String>>,
    pub auto_verified_attributes: Option<Vec<String>>,
    pub device_configuration: Option<DeviceConfigurationType>,
    pub email_configuration: Option<EmailConfigurationType>,
    pub email_verification_message: Option<String>,
    pub email_verification_subject: Option<String>,
    pub lambda_config: Option<LambdaConfigType>,
    pub mfa_configuration: Option<String>,
    pub policies: Option<UserPoolPolicyType>,
    pub pool_name: String,
    pub schema: Option<Vec<SchemaAttributeType>>,
    pub sms_authentication_message: Option<String>,
    pub sms_configuration: Option<SmsConfigurationType>,
    pub sms_verification_message: Option<String>,
    pub user_pool_tags: Option<HashMap<String, String>>,
}

Represents the request to create a user pool.

Fields
admin_create_user_config: Option<AdminCreateUserConfigType>

The configuration for AdminCreateUser requests.

alias_attributes: Option<Vec<String>>

Attributes supported as an alias for this user pool. Possible values: phone_number, email, or preferred_username.

auto_verified_attributes: Option<Vec<String>>

The attributes to be auto-verified. Possible values: email, phone_number.

device_configuration: Option<DeviceConfigurationType>

The device configuration.

email_configuration: Option<EmailConfigurationType>

The email configuration.

email_verification_message: Option<String>

A string representing the email verification message.

email_verification_subject: Option<String>

A string representing the email verification subject.

lambda_config: Option<LambdaConfigType>

The Lambda trigger configuration information for the new user pool.

mfa_configuration: Option<String>

Specifies MFA configuration details.

policies: Option<UserPoolPolicyType>

The policies associated with the new user pool.

pool_name: String

A string used to name the user pool.

schema: Option<Vec<SchemaAttributeType>>

An array of schema attributes for the new user pool. These attributes can be standard or custom attributes.

sms_authentication_message: Option<String>

A string representing the SMS authentication message.

sms_configuration: Option<SmsConfigurationType>

The SMS configuration.

sms_verification_message: Option<String>

A string representing the SMS verification message.

user_pool_tags: Option<HashMap<String, String>>

The cost allocation tags for the user pool. For more information, see Adding Cost Allocation Tags to Your User Pool