Struct rusoto_cognito_idp::ListUsersRequest

pub struct ListUsersRequest {
    pub attributes_to_get: Option<Vec<String>>,
    pub filter: Option<String>,
    pub limit: Option<i64>,
    pub pagination_token: Option<String>,
    pub user_pool_id: String,
}

Represents the request to list users.

Fields
attributes_to_get: Option<Vec<String>>

An array of strings, where each string is the name of a user attribute to be returned for each user in the search results. If the array is empty, all attributes are returned.

filter: Option<String>

A filter string of the form "AttributeName Filter-Type "AttributeValue"". Quotation marks within the filter string must be escaped using the backslash (\) character. For example, "family_name = \"Reddy\"".

AttributeName: The name of the attribute to search for. You can only search for one attribute at a time.

Filter-Type: For an exact match, use =, for example, "given_name = \"Jon\"". For a prefix ("starts with") match, use ^=, for example, "given_name ^= \"Jon\"".

AttributeValue: The attribute value that must be matched for each user.

If the filter string is empty, ListUsers returns all users in the user pool.

You can only search for the following standard attributes:

username (case-sensitive)

email

phone_number

name

given_name

family_name

preferred_username

cognito:user_status (called Enabled in the Console) (case-sensitive)

status (case-insensitive)

Custom attributes are not searchable.

For more information, see Searching for Users Using the ListUsers API and Examples of Using the ListUsers API in the Amazon Cognito Developer Guide.

limit: Option<i64>

Maximum number of users to be returned.

pagination_token: Option<String>

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

user_pool_id: String

The user pool ID for the user pool on which the search should be performed.