Struct rusoto_cognito_idp::RiskConfigurationType

pub struct RiskConfigurationType {
    pub account_takeover_risk_configuration: Option<AccountTakeoverRiskConfigurationType>,
    pub client_id: Option<String>,
    pub compromised_credentials_risk_configuration: Option<CompromisedCredentialsRiskConfigurationType>,
    pub last_modified_date: Option<f64>,
    pub risk_exception_configuration: Option<RiskExceptionConfigurationType>,
    pub user_pool_id: Option<String>,
}

The risk configuration type.

Fields
account_takeover_risk_configuration: Option<AccountTakeoverRiskConfigurationType>

The account takeover risk configuration object including the NotifyConfiguration object and Actions to take in the case of an account takeover.

client_id: Option<String>

The app client ID.

compromised_credentials_risk_configuration: Option<CompromisedCredentialsRiskConfigurationType>

The compromised credentials risk configuration object including the EventFilter and the EventAction

last_modified_date: Option<f64>

The last modified date.

risk_exception_configuration: Option<RiskExceptionConfigurationType>

The configuration to override the risk decision.

user_pool_id: Option<String>

The user pool ID.