Struct rusoto_cognito_idp::UserImportJobType

pub struct UserImportJobType {
    pub cloud_watch_logs_role_arn: Option<String>,
    pub completion_date: Option<f64>,
    pub completion_message: Option<String>,
    pub creation_date: Option<f64>,
    pub failed_users: Option<i64>,
    pub imported_users: Option<i64>,
    pub job_id: Option<String>,
    pub job_name: Option<String>,
    pub pre_signed_url: Option<String>,
    pub skipped_users: Option<i64>,
    pub start_date: Option<f64>,
    pub status: Option<String>,
    pub user_pool_id: Option<String>,
}

The user import job type.

Fields
cloud_watch_logs_role_arn: Option<String>

The role ARN for the Amazon CloudWatch Logging role for the user import job. For more information, see "Creating the CloudWatch Logs IAM Role" in the Amazon Cognito Developer Guide.

completion_date: Option<f64>

The date when the user import job was completed.

completion_message: Option<String>

The message returned when the user import job is completed.

creation_date: Option<f64>

The date the user import job was created.

failed_users: Option<i64>

The number of users that could not be imported.

imported_users: Option<i64>

The number of users that were successfully imported.

job_id: Option<String>

The job ID for the user import job.

job_name: Option<String>

The job name for the user import job.

pre_signed_url: Option<String>

The pre-signed URL to be used to upload the .csv file.

skipped_users: Option<i64>

The number of users that were skipped.

start_date: Option<f64>

The date when the user import job was started.

status: Option<String>

The status of the user import job. One of the following:

Created - The job was created but not started.

Pending - A transition state. You have started the job, but it has not begun importing users yet.

InProgress - The job has started, and users are being imported.

Stopping - You have stopped the job, but the job has not stopped importing users yet.

Stopped - You have stopped the job, and the job has stopped importing users.

Succeeded - The job has completed successfully.

Failed - The job has stopped due to an error.

Expired - You created a job, but did not start the job within 24-48 hours. All data associated with the job was deleted, and the job cannot be started.

user_pool_id: Option<String>

The user pool ID for the user pool that the users are being imported into.