Struct rusoto_cognito_idp::VerificationMessageTemplateType

pub struct VerificationMessageTemplateType {
    pub default_email_option: Option<String>,
    pub email_message: Option<String>,
    pub email_message_by_link: Option<String>,
    pub email_subject: Option<String>,
    pub email_subject_by_link: Option<String>,
    pub sms_message: Option<String>,
}

The template for verification messages.

Fields
default_email_option: Option<String>

The default email option.

email_message: Option<String>

The email message template.

email_message_by_link: Option<String>

The email message template for sending a confirmation link to the user.

email_subject: Option<String>

The subject line for the email message template.

email_subject_by_link: Option<String>

The subject line for the email message template for sending a confirmation link to the user.

sms_message: Option<String>

The SMS message template.