Crate rusoto_cognito_idp

Using the Amazon Cognito User Pools API, you can create a user pool to manage directories and users. You can authenticate a user to obtain tokens related to user identity and access policies.

This API reference provides information about user pools in Amazon Cognito User Pools.

For more information, see the Amazon Cognito Documentation.

If you're using the service, you're probably looking for CognitoIdentityProviderClient and CognitoIdentityProvider.

Structs

AddCustomAttributesRequest
Represents the request to add custom attributes.

AddCustomAttributesResponse
Represents the response from the server for the request to add custom attributes.

AdminAddUserToGroupRequest

AdminConfirmSignUpRequest
Represents the request to confirm user registration.

AdminConfirmSignUpResponse
Represents the response from the server for the request to confirm registration.

AdminCreateUserConfigType
The type of configuration for creating a new user profile.

AdminCreateUserRequest
Represents the request to create a user in the specified user pool.

AdminCreateUserResponse
Represents the response from the server to the request to create the user.

AdminDeleteUserAttributesRequest
Represents the request to delete user attributes as an administrator.

AdminDeleteUserAttributesResponse
Represents the response received from the server for a request to delete user attributes.

AdminDeleteUserRequest
Represents the request to delete a user as an administrator.

AdminDisableUserRequest
Represents the request to disable any user as an administrator.

AdminDisableUserResponse
Represents the response received from the server to disable the user as an administrator.

AdminEnableUserRequest
Represents the request that enables the user as an administrator.

AdminEnableUserResponse
Represents the response from the server for the request to enable a user as an administrator.

AdminForgetDeviceRequest
Sends the forgot device request, as an administrator.

AdminGetDeviceRequest
Represents the request to get the device, as an administrator.

AdminGetDeviceResponse
Gets the device response, as an administrator.

AdminGetUserRequest
Represents the request to get the specified user as an administrator.

AdminGetUserResponse
Represents the response from the server from the request to get the specified user as an administrator.

AdminInitiateAuthRequest
Initiates the authorization request, as an administrator.

AdminInitiateAuthResponse
Initiates the authentication response, as an administrator.

AdminListDevicesRequest
Represents the request to list devices, as an administrator.

AdminListDevicesResponse
Lists the device's response, as an administrator.

AdminListGroupsForUserRequest

AdminListGroupsForUserResponse

AdminRemoveUserFromGroupRequest

AdminResetUserPasswordRequest
Represents the request to reset a user's password as an administrator.

AdminResetUserPasswordResponse
Represents the response from the server to reset a user password as an administrator.

AdminRespondToAuthChallengeRequest
The request to respond to the authentication challenge, as an administrator.

AdminRespondToAuthChallengeResponse
Responds to the authentication challenge, as an administrator.

AdminSetUserSettingsRequest
Represents the request to set user settings as an administrator.

AdminSetUserSettingsResponse
Represents the response from the server to set user settings as an administrator.

AdminUpdateDeviceStatusRequest
The request to update the device status, as an administrator.

AdminUpdateDeviceStatusResponse
The status response from the request to update the device, as an administrator.

AdminUpdateUserAttributesRequest
Represents the request to update the user's attributes as an administrator.

AdminUpdateUserAttributesResponse
Represents the response from the server for the request to update user attributes as an administrator.

AdminUserGlobalSignOutRequest
The request to sign out of all devices, as an administrator.

AdminUserGlobalSignOutResponse
The global sign-out response, as an administrator.

AttributeType
Specifies whether the attribute is standard or custom.

AuthenticationResultType
The result type of the authentication result.

ChangePasswordRequest
Represents the request to change a user password.

ChangePasswordResponse
The response from the server to the change password request.

CodeDeliveryDetailsType
The type of code delivery details being returned from the server.

CognitoIdentityProviderClient
A client for the Amazon Cognito Identity Provider API.

ConfirmDeviceRequest
Confirms the device request.

ConfirmDeviceResponse
Confirms the device response.

ConfirmForgotPasswordRequest
The request representing the confirmation for a password reset.

ConfirmForgotPasswordResponse
The response from the server that results from a user's request to retrieve a forgotten password.

ConfirmSignUpRequest
Represents the request to confirm registration of a user.

ConfirmSignUpResponse
Represents the response from the server for the registration confirmation.

CreateGroupRequest

CreateGroupResponse

CreateIdentityProviderRequest

CreateIdentityProviderResponse

CreateUserImportJobRequest
Represents the request to create the user import job.

CreateUserImportJobResponse
Represents the response from the server to the request to create the user import job.

CreateUserPoolClientRequest
Represents the request to create a user pool client.

CreateUserPoolClientResponse
Represents the response from the server to create a user pool client.

CreateUserPoolDomainRequest

CreateUserPoolDomainResponse

CreateUserPoolRequest
Represents the request to create a user pool.

CreateUserPoolResponse
Represents the response from the server for the request to create a user pool.

DeleteGroupRequest

DeleteIdentityProviderRequest

DeleteUserAttributesRequest
Represents the request to delete user attributes.

DeleteUserAttributesResponse
Represents the response from the server to delete user attributes.

DeleteUserPoolClientRequest
Represents the request to delete a user pool client.

DeleteUserPoolDomainRequest

DeleteUserPoolDomainResponse

DeleteUserPoolRequest
Represents the request to delete a user pool.

DeleteUserRequest
Represents the request to delete a user.

DescribeIdentityProviderRequest

DescribeIdentityProviderResponse

DescribeUserImportJobRequest
Represents the request to describe the user import job.

DescribeUserImportJobResponse
Represents the response from the server to the request to describe the user import job.

DescribeUserPoolClientRequest
Represents the request to describe a user pool client.

DescribeUserPoolClientResponse
Represents the response from the server from a request to describe the user pool client.

DescribeUserPoolDomainRequest

DescribeUserPoolDomainResponse

DescribeUserPoolRequest
Represents the request to describe the user pool.

DescribeUserPoolResponse
Represents the response to describe the user pool.

DeviceConfigurationType
The type of configuration for the user pool's device tracking.

DeviceSecretVerifierConfigType
The device verifier against which it will be authenticated.

DeviceType
The device type.

DomainDescriptionType
A container for information about a domain.

EmailConfigurationType
The email configuration type.

ForgetDeviceRequest
Represents the request to forget the device.

ForgotPasswordRequest
Represents the request to reset a user's password.

ForgotPasswordResponse
Respresents the response from the server regarding the request to reset a password.

GetCSVHeaderRequest
Represents the request to get the header information for the .csv file for the user import job.

GetCSVHeaderResponse
Represents the response from the server to the request to get the header information for the .csv file for the user import job.

GetDeviceRequest
Represents the request to get the device.

GetDeviceResponse
Gets the device response.

GetGroupRequest

GetGroupResponse

GetIdentityProviderByIdentifierRequest

GetIdentityProviderByIdentifierResponse

GetUserAttributeVerificationCodeRequest
Represents the request to get user attribute verification.

GetUserAttributeVerificationCodeResponse
The verification code response returned by the server response to get the user attribute verification code.

GetUserRequest
Represents the request to get information about the user.

GetUserResponse
Represents the response from the server from the request to get information about the user.

GlobalSignOutRequest
Represents the request to sign out all devices.

GlobalSignOutResponse
The response to the request to sign out all devices.

GroupType
The group type.

IdentityProviderType
A container for information about an identity provider.

InitiateAuthRequest
Initiates the authentication request.

InitiateAuthResponse
Initiates the authentication response.

LambdaConfigType
Specifies the type of configuration for AWS Lambda triggers.

ListDevicesRequest
Represents the request to list the devices.

ListDevicesResponse
Represents the response to list devices.

ListGroupsRequest

ListGroupsResponse

ListIdentityProvidersRequest

ListIdentityProvidersResponse

ListUserImportJobsRequest
Represents the request to list the user import jobs.

ListUserImportJobsResponse
Represents the response from the server to the request to list the user import jobs.

ListUserPoolClientsRequest
Represents the request to list the user pool clients.

ListUserPoolClientsResponse
Represents the response from the server that lists user pool clients.

ListUserPoolsRequest
Represents the request to list user pools.

ListUserPoolsResponse
Represents the response to list user pools.

ListUsersInGroupRequest

ListUsersInGroupResponse

ListUsersRequest
Represents the request to list users.

ListUsersResponse
The response from the request to list users.

MFAOptionType
Specifies the different settings for multi-factor authentication (MFA).

MessageTemplateType
The message template structure.

NewDeviceMetadataType
The new device metadata type.

NumberAttributeConstraintsType
The minimum and maximum value of an attribute that is of the number data type.

PasswordPolicyType
The password policy type.

ProviderDescription
A container for identity provider details.

ResendConfirmationCodeRequest
Represents the request to resend the confirmation code.

ResendConfirmationCodeResponse
The response from the server when the Amazon Cognito Your User Pools service makes the request to resend a confirmation code.

RespondToAuthChallengeRequest
The request to respond to an authentication challenge.

RespondToAuthChallengeResponse
The response to respond to the authentication challenge.

SchemaAttributeType
Contains information about the schema attribute.

SetUserSettingsRequest
Represents the request to set user settings.

SetUserSettingsResponse
The response from the server for a set user settings request.

SignUpRequest
Represents the request to register a user.

SignUpResponse
The response from the server for a registration request.

SmsConfigurationType
The SMS configuration type.

StartUserImportJobRequest
Represents the request to start the user import job.

StartUserImportJobResponse
Represents the response from the server to the request to start the user import job.

StopUserImportJobRequest
Represents the request to stop the user import job.

StopUserImportJobResponse
Represents the response from the server to the request to stop the user import job.

StringAttributeConstraintsType
The type of constraints associated with an attribute of the string type.

UpdateDeviceStatusRequest
Represents the request to update the device status.

UpdateDeviceStatusResponse
The response to the request to update the device status.

UpdateGroupRequest

UpdateGroupResponse

UpdateIdentityProviderRequest

UpdateIdentityProviderResponse

UpdateUserAttributesRequest
Represents the request to update user attributes.

UpdateUserAttributesResponse
Represents the response from the server for the request to update user attributes.

UpdateUserPoolClientRequest
Represents the request to update the user pool client.

UpdateUserPoolClientResponse
Represents the response from the server to the request to update the user pool client.

UpdateUserPoolRequest
Represents the request to update the user pool.

UpdateUserPoolResponse
Represents the response from the server when you make a request to update the user pool.

UserImportJobType
The user import job type.

UserPoolClientDescription
The description of the user pool client.

UserPoolClientType
A user pool of the client type.

UserPoolDescriptionType
A user pool description.

UserPoolPolicyType
The type of policy in a user pool.

UserPoolType
A container for information about the user pool type.

UserType
The user type.

VerifyUserAttributeRequest
Represents the request to verify user attributes.

VerifyUserAttributeResponse
A container representing the response from the server from the request to verify user attributes.

Enums

AddCustomAttributesError
Errors returned by AddCustomAttributes

AdminAddUserToGroupError
Errors returned by AdminAddUserToGroup

AdminConfirmSignUpError
Errors returned by AdminConfirmSignUp

AdminCreateUserError
Errors returned by AdminCreateUser

AdminDeleteUserAttributesError
Errors returned by AdminDeleteUserAttributes

AdminDeleteUserError
Errors returned by AdminDeleteUser

AdminDisableUserError
Errors returned by AdminDisableUser

AdminEnableUserError
Errors returned by AdminEnableUser

AdminForgetDeviceError
Errors returned by AdminForgetDevice

AdminGetDeviceError
Errors returned by AdminGetDevice

AdminGetUserError
Errors returned by AdminGetUser

AdminInitiateAuthError
Errors returned by AdminInitiateAuth

AdminListDevicesError
Errors returned by AdminListDevices

AdminListGroupsForUserError
Errors returned by AdminListGroupsForUser

AdminRemoveUserFromGroupError
Errors returned by AdminRemoveUserFromGroup

AdminResetUserPasswordError
Errors returned by AdminResetUserPassword

AdminRespondToAuthChallengeError
Errors returned by AdminRespondToAuthChallenge

AdminSetUserSettingsError
Errors returned by AdminSetUserSettings

AdminUpdateDeviceStatusError
Errors returned by AdminUpdateDeviceStatus

AdminUpdateUserAttributesError
Errors returned by AdminUpdateUserAttributes

AdminUserGlobalSignOutError
Errors returned by AdminUserGlobalSignOut

ChangePasswordError
Errors returned by ChangePassword

ConfirmDeviceError
Errors returned by ConfirmDevice

ConfirmForgotPasswordError
Errors returned by ConfirmForgotPassword

ConfirmSignUpError
Errors returned by ConfirmSignUp

CreateGroupError
Errors returned by CreateGroup

CreateIdentityProviderError
Errors returned by CreateIdentityProvider

CreateUserImportJobError
Errors returned by CreateUserImportJob

CreateUserPoolClientError
Errors returned by CreateUserPoolClient

CreateUserPoolDomainError
Errors returned by CreateUserPoolDomain

CreateUserPoolError
Errors returned by CreateUserPool

DeleteGroupError
Errors returned by DeleteGroup

DeleteIdentityProviderError
Errors returned by DeleteIdentityProvider

DeleteUserAttributesError
Errors returned by DeleteUserAttributes

DeleteUserError
Errors returned by DeleteUser

DeleteUserPoolClientError
Errors returned by DeleteUserPoolClient

DeleteUserPoolDomainError
Errors returned by DeleteUserPoolDomain

DeleteUserPoolError
Errors returned by DeleteUserPool

DescribeIdentityProviderError
Errors returned by DescribeIdentityProvider

DescribeUserImportJobError
Errors returned by DescribeUserImportJob

DescribeUserPoolClientError
Errors returned by DescribeUserPoolClient

DescribeUserPoolDomainError
Errors returned by DescribeUserPoolDomain

DescribeUserPoolError
Errors returned by DescribeUserPool

ForgetDeviceError
Errors returned by ForgetDevice

ForgotPasswordError
Errors returned by ForgotPassword

GetCSVHeaderError
Errors returned by GetCSVHeader

GetDeviceError
Errors returned by GetDevice

GetGroupError
Errors returned by GetGroup

GetIdentityProviderByIdentifierError
Errors returned by GetIdentityProviderByIdentifier

GetUserAttributeVerificationCodeError
Errors returned by GetUserAttributeVerificationCode

GetUserError
Errors returned by GetUser

GlobalSignOutError
Errors returned by GlobalSignOut

InitiateAuthError
Errors returned by InitiateAuth

ListDevicesError
Errors returned by ListDevices

ListGroupsError
Errors returned by ListGroups

ListIdentityProvidersError
Errors returned by ListIdentityProviders

ListUserImportJobsError
Errors returned by ListUserImportJobs

ListUserPoolClientsError
Errors returned by ListUserPoolClients

ListUserPoolsError
Errors returned by ListUserPools

ListUsersError
Errors returned by ListUsers

ListUsersInGroupError
Errors returned by ListUsersInGroup

ResendConfirmationCodeError
Errors returned by ResendConfirmationCode

RespondToAuthChallengeError
Errors returned by RespondToAuthChallenge

SetUserSettingsError
Errors returned by SetUserSettings

SignUpError
Errors returned by SignUp

StartUserImportJobError
Errors returned by StartUserImportJob

StopUserImportJobError
Errors returned by StopUserImportJob

UpdateDeviceStatusError
Errors returned by UpdateDeviceStatus

UpdateGroupError
Errors returned by UpdateGroup

UpdateIdentityProviderError
Errors returned by UpdateIdentityProvider

UpdateUserAttributesError
Errors returned by UpdateUserAttributes

UpdateUserPoolClientError
Errors returned by UpdateUserPoolClient

UpdateUserPoolError
Errors returned by UpdateUserPool

VerifyUserAttributeError
Errors returned by VerifyUserAttribute

Traits

CognitoIdentityProvider
Trait representing the capabilities of the Amazon Cognito Identity Provider API. Amazon Cognito Identity Provider clients implement this trait.