Crate rusoto_cognito_idp

Using the Amazon Cognito User Pools API, you can create a user pool to manage directories and users. You can authenticate a user to obtain tokens related to user identity and access policies.

This API reference provides information about user pools in Amazon Cognito User Pools.

For more information, see the Amazon Cognito Documentation.

If you're using the service, you're probably looking for CognitoIdentityProviderClient and CognitoIdentityProvider.

Structs

AccountTakeoverActionType
Account takeover action type.

AccountTakeoverActionsType
Account takeover actions type.

AccountTakeoverRiskConfigurationType
Configuration for mitigation actions and notification for different levels of risk detected for a potential account takeover.

AddCustomAttributesRequest
Represents the request to add custom attributes.

AddCustomAttributesResponse
Represents the response from the server for the request to add custom attributes.

AdminAddUserToGroupRequest

AdminConfirmSignUpRequest
Represents the request to confirm user registration.

AdminConfirmSignUpResponse
Represents the response from the server for the request to confirm registration.

AdminCreateUserConfigType
The configuration for creating a new user profile.

AdminCreateUserRequest
Represents the request to create a user in the specified user pool.

AdminCreateUserResponse
Represents the response from the server to the request to create the user.

AdminDeleteUserAttributesRequest
Represents the request to delete user attributes as an administrator.

AdminDeleteUserAttributesResponse
Represents the response received from the server for a request to delete user attributes.

AdminDeleteUserRequest
Represents the request to delete a user as an administrator.

AdminDisableProviderForUserRequest

AdminDisableProviderForUserResponse

AdminDisableUserRequest
Represents the request to disable any user as an administrator.

AdminDisableUserResponse
Represents the response received from the server to disable the user as an administrator.

AdminEnableUserRequest
Represents the request that enables the user as an administrator.

AdminEnableUserResponse
Represents the response from the server for the request to enable a user as an administrator.

AdminForgetDeviceRequest
Sends the forgot device request, as an administrator.

AdminGetDeviceRequest
Represents the request to get the device, as an administrator.

AdminGetDeviceResponse
Gets the device response, as an administrator.

AdminGetUserRequest
Represents the request to get the specified user as an administrator.

AdminGetUserResponse
Represents the response from the server from the request to get the specified user as an administrator.

AdminInitiateAuthRequest
Initiates the authorization request, as an administrator.

AdminInitiateAuthResponse
Initiates the authentication response, as an administrator.

AdminLinkProviderForUserRequest

AdminLinkProviderForUserResponse

AdminListDevicesRequest
Represents the request to list devices, as an administrator.

AdminListDevicesResponse
Lists the device's response, as an administrator.

AdminListGroupsForUserRequest

AdminListGroupsForUserResponse

AdminListUserAuthEventsRequest

AdminListUserAuthEventsResponse

AdminRemoveUserFromGroupRequest

AdminResetUserPasswordRequest
Represents the request to reset a user's password as an administrator.

AdminResetUserPasswordResponse
Represents the response from the server to reset a user password as an administrator.

AdminRespondToAuthChallengeRequest
The request to respond to the authentication challenge, as an administrator.

AdminRespondToAuthChallengeResponse
Responds to the authentication challenge, as an administrator.

AdminSetUserMFAPreferenceRequest

AdminSetUserMFAPreferenceResponse

AdminSetUserSettingsRequest
Represents the request to set user settings as an administrator.

AdminSetUserSettingsResponse
Represents the response from the server to set user settings as an administrator.

AdminUpdateAuthEventFeedbackRequest

AdminUpdateAuthEventFeedbackResponse

AdminUpdateDeviceStatusRequest
The request to update the device status, as an administrator.

AdminUpdateDeviceStatusResponse
The status response from the request to update the device, as an administrator.

AdminUpdateUserAttributesRequest
Represents the request to update the user's attributes as an administrator.

AdminUpdateUserAttributesResponse
Represents the response from the server for the request to update user attributes as an administrator.

AdminUserGlobalSignOutRequest
The request to sign out of all devices, as an administrator.

AdminUserGlobalSignOutResponse
The global sign-out response, as an administrator.

AnalyticsConfigurationType
The Amazon Pinpoint analytics configuration for collecting metrics for a user pool.

AnalyticsMetadataType
An Amazon Pinpoint analytics endpoint.

An endpoint uniquely identifies a mobile device, email address, or phone number that can receive messages from Amazon Pinpoint analytics.

AssociateSoftwareTokenRequest

AssociateSoftwareTokenResponse

AttributeType
Specifies whether the attribute is standard or custom.

AuthEventType
The authentication event type.

AuthenticationResultType
The authentication result.

ChallengeResponseType
The challenge response type.

ChangePasswordRequest
Represents the request to change a user password.

ChangePasswordResponse
The response from the server to the change password request.

CodeDeliveryDetailsType
The code delivery details being returned from the server.

CognitoIdentityProviderClient
A client for the Amazon Cognito Identity Provider API.

CompromisedCredentialsActionsType
The compromised credentials actions type

CompromisedCredentialsRiskConfigurationType
The compromised credentials risk configuration type.

ConfirmDeviceRequest
Confirms the device request.

ConfirmDeviceResponse
Confirms the device response.

ConfirmForgotPasswordRequest
The request representing the confirmation for a password reset.

ConfirmForgotPasswordResponse
The response from the server that results from a user's request to retrieve a forgotten password.

ConfirmSignUpRequest
Represents the request to confirm registration of a user.

ConfirmSignUpResponse
Represents the response from the server for the registration confirmation.

ContextDataType
Contextual user data type used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

CreateGroupRequest

CreateGroupResponse

CreateIdentityProviderRequest

CreateIdentityProviderResponse

CreateResourceServerRequest

CreateResourceServerResponse

CreateUserImportJobRequest
Represents the request to create the user import job.

CreateUserImportJobResponse
Represents the response from the server to the request to create the user import job.

CreateUserPoolClientRequest
Represents the request to create a user pool client.

CreateUserPoolClientResponse
Represents the response from the server to create a user pool client.

CreateUserPoolDomainRequest

CreateUserPoolDomainResponse

CreateUserPoolRequest
Represents the request to create a user pool.

CreateUserPoolResponse
Represents the response from the server for the request to create a user pool.

DeleteGroupRequest

DeleteIdentityProviderRequest

DeleteResourceServerRequest

DeleteUserAttributesRequest
Represents the request to delete user attributes.

DeleteUserAttributesResponse
Represents the response from the server to delete user attributes.

DeleteUserPoolClientRequest
Represents the request to delete a user pool client.

DeleteUserPoolDomainRequest

DeleteUserPoolDomainResponse

DeleteUserPoolRequest
Represents the request to delete a user pool.

DeleteUserRequest
Represents the request to delete a user.

DescribeIdentityProviderRequest

DescribeIdentityProviderResponse

DescribeResourceServerRequest

DescribeResourceServerResponse

DescribeRiskConfigurationRequest

DescribeRiskConfigurationResponse

DescribeUserImportJobRequest
Represents the request to describe the user import job.

DescribeUserImportJobResponse
Represents the response from the server to the request to describe the user import job.

DescribeUserPoolClientRequest
Represents the request to describe a user pool client.

DescribeUserPoolClientResponse
Represents the response from the server from a request to describe the user pool client.

DescribeUserPoolDomainRequest

DescribeUserPoolDomainResponse

DescribeUserPoolRequest
Represents the request to describe the user pool.

DescribeUserPoolResponse
Represents the response to describe the user pool.

DeviceConfigurationType
The configuration for the user pool's device tracking.

DeviceSecretVerifierConfigType
The device verifier against which it will be authenticated.

DeviceType
The device type.

DomainDescriptionType
A container for information about a domain.

EmailConfigurationType
The email configuration type.

EventContextDataType
Specifies the user context data captured at the time of an event request.

EventFeedbackType
Specifies the event feedback type.

EventRiskType
The event risk type.

ForgetDeviceRequest
Represents the request to forget the device.

ForgotPasswordRequest
Represents the request to reset a user's password.

ForgotPasswordResponse
Respresents the response from the server regarding the request to reset a password.

GetCSVHeaderRequest
Represents the request to get the header information for the .csv file for the user import job.

GetCSVHeaderResponse
Represents the response from the server to the request to get the header information for the .csv file for the user import job.

GetDeviceRequest
Represents the request to get the device.

GetDeviceResponse
Gets the device response.

GetGroupRequest

GetGroupResponse

GetIdentityProviderByIdentifierRequest

GetIdentityProviderByIdentifierResponse

GetSigningCertificateRequest
Request to get a signing certificate from Cognito.

GetSigningCertificateResponse
Response from Cognito for a signing certificate request.

GetUICustomizationRequest

GetUICustomizationResponse

GetUserAttributeVerificationCodeRequest
Represents the request to get user attribute verification.

GetUserAttributeVerificationCodeResponse
The verification code response returned by the server response to get the user attribute verification code.

GetUserPoolMfaConfigRequest

GetUserPoolMfaConfigResponse

GetUserRequest
Represents the request to get information about the user.

GetUserResponse
Represents the response from the server from the request to get information about the user.

GlobalSignOutRequest
Represents the request to sign out all devices.

GlobalSignOutResponse
The response to the request to sign out all devices.

GroupType
The group type.

HttpHeader
The HTTP header.

IdentityProviderType
A container for information about an identity provider.

InitiateAuthRequest
Initiates the authentication request.

InitiateAuthResponse
Initiates the authentication response.

LambdaConfigType
Specifies the configuration for AWS Lambda triggers.

ListDevicesRequest
Represents the request to list the devices.

ListDevicesResponse
Represents the response to list devices.

ListGroupsRequest

ListGroupsResponse

ListIdentityProvidersRequest

ListIdentityProvidersResponse

ListResourceServersRequest

ListResourceServersResponse

ListUserImportJobsRequest
Represents the request to list the user import jobs.

ListUserImportJobsResponse
Represents the response from the server to the request to list the user import jobs.

ListUserPoolClientsRequest
Represents the request to list the user pool clients.

ListUserPoolClientsResponse
Represents the response from the server that lists user pool clients.

ListUserPoolsRequest
Represents the request to list user pools.

ListUserPoolsResponse
Represents the response to list user pools.

ListUsersInGroupRequest

ListUsersInGroupResponse

ListUsersRequest
Represents the request to list users.

ListUsersResponse
The response from the request to list users.

MFAOptionType
Specifies the different settings for multi-factor authentication (MFA).

MessageTemplateType
The message template structure.

NewDeviceMetadataType
The new device metadata type.

NotifyConfigurationType
The notify configuration type.

NotifyEmailType
The notify email type.

NumberAttributeConstraintsType
The minimum and maximum value of an attribute that is of the number data type.

PasswordPolicyType
The password policy type.

ProviderDescription
A container for identity provider details.

ProviderUserIdentifierType
A container for information about an identity provider for a user pool.

ResendConfirmationCodeRequest
Represents the request to resend the confirmation code.

ResendConfirmationCodeResponse
The response from the server when the Amazon Cognito Your User Pools service makes the request to resend a confirmation code.

ResourceServerScopeType
A resource server scope.

ResourceServerType
A container for information about a resource server for a user pool.

RespondToAuthChallengeRequest
The request to respond to an authentication challenge.

RespondToAuthChallengeResponse
The response to respond to the authentication challenge.

RiskConfigurationType
The risk configuration type.

RiskExceptionConfigurationType
The type of the configuration to override the risk decision.

SMSMfaSettingsType
The SMS multi-factor authentication (MFA) settings type.

SchemaAttributeType
Contains information about the schema attribute.

SetRiskConfigurationRequest

SetRiskConfigurationResponse

SetUICustomizationRequest

SetUICustomizationResponse

SetUserMFAPreferenceRequest

SetUserMFAPreferenceResponse

SetUserPoolMfaConfigRequest

SetUserPoolMfaConfigResponse

SetUserSettingsRequest
Represents the request to set user settings.

SetUserSettingsResponse
The response from the server for a set user settings request.

SignUpRequest
Represents the request to register a user.

SignUpResponse
The response from the server for a registration request.

SmsConfigurationType
The SMS configuration type.

SmsMfaConfigType
The SMS text message multi-factor authentication (MFA) configuration type.

SoftwareTokenMfaConfigType
The type used for enabling software token MFA at the user pool level.

SoftwareTokenMfaSettingsType
The type used for enabling software token MFA at the user level.

StartUserImportJobRequest
Represents the request to start the user import job.

StartUserImportJobResponse
Represents the response from the server to the request to start the user import job.

StopUserImportJobRequest
Represents the request to stop the user import job.

StopUserImportJobResponse
Represents the response from the server to the request to stop the user import job.

StringAttributeConstraintsType
The constraints associated with a string attribute.

UICustomizationType
A container for the UI customization information for a user pool's built-in app UI.

UpdateAuthEventFeedbackRequest

UpdateAuthEventFeedbackResponse

UpdateDeviceStatusRequest
Represents the request to update the device status.

UpdateDeviceStatusResponse
The response to the request to update the device status.

UpdateGroupRequest

UpdateGroupResponse

UpdateIdentityProviderRequest

UpdateIdentityProviderResponse

UpdateResourceServerRequest

UpdateResourceServerResponse

UpdateUserAttributesRequest
Represents the request to update user attributes.

UpdateUserAttributesResponse
Represents the response from the server for the request to update user attributes.

UpdateUserPoolClientRequest
Represents the request to update the user pool client.

UpdateUserPoolClientResponse
Represents the response from the server to the request to update the user pool client.

UpdateUserPoolRequest
Represents the request to update the user pool.

UpdateUserPoolResponse
Represents the response from the server when you make a request to update the user pool.

UserContextDataType
Contextual data such as the user's device fingerprint, IP address, or location used for evaluating the risk of an unexpected event by Amazon Cognito advanced security.

UserImportJobType
The user import job type.

UserPoolAddOnsType
The user pool add-ons type.

UserPoolClientDescription
The description of the user pool client.

UserPoolClientType
Contains information about a user pool client.

UserPoolDescriptionType
A user pool description.

UserPoolPolicyType
The policy associated with a user pool.

UserPoolType
A container for information about the user pool.

UserType
The user type.

VerificationMessageTemplateType
The template for verification messages.

VerifySoftwareTokenRequest

VerifySoftwareTokenResponse

VerifyUserAttributeRequest
Represents the request to verify user attributes.

VerifyUserAttributeResponse
A container representing the response from the server from the request to verify user attributes.

Enums

AddCustomAttributesError
Errors returned by AddCustomAttributes

AdminAddUserToGroupError
Errors returned by AdminAddUserToGroup

AdminConfirmSignUpError
Errors returned by AdminConfirmSignUp

AdminCreateUserError
Errors returned by AdminCreateUser

AdminDeleteUserAttributesError
Errors returned by AdminDeleteUserAttributes

AdminDeleteUserError
Errors returned by AdminDeleteUser

AdminDisableProviderForUserError
Errors returned by AdminDisableProviderForUser

AdminDisableUserError
Errors returned by AdminDisableUser

AdminEnableUserError
Errors returned by AdminEnableUser

AdminForgetDeviceError
Errors returned by AdminForgetDevice

AdminGetDeviceError
Errors returned by AdminGetDevice

AdminGetUserError
Errors returned by AdminGetUser

AdminInitiateAuthError
Errors returned by AdminInitiateAuth

AdminLinkProviderForUserError
Errors returned by AdminLinkProviderForUser

AdminListDevicesError
Errors returned by AdminListDevices

AdminListGroupsForUserError
Errors returned by AdminListGroupsForUser

AdminListUserAuthEventsError
Errors returned by AdminListUserAuthEvents

AdminRemoveUserFromGroupError
Errors returned by AdminRemoveUserFromGroup

AdminResetUserPasswordError
Errors returned by AdminResetUserPassword

AdminRespondToAuthChallengeError
Errors returned by AdminRespondToAuthChallenge

AdminSetUserMFAPreferenceError
Errors returned by AdminSetUserMFAPreference

AdminSetUserSettingsError
Errors returned by AdminSetUserSettings

AdminUpdateAuthEventFeedbackError
Errors returned by AdminUpdateAuthEventFeedback

AdminUpdateDeviceStatusError
Errors returned by AdminUpdateDeviceStatus

AdminUpdateUserAttributesError
Errors returned by AdminUpdateUserAttributes

AdminUserGlobalSignOutError
Errors returned by AdminUserGlobalSignOut

AssociateSoftwareTokenError
Errors returned by AssociateSoftwareToken

ChangePasswordError
Errors returned by ChangePassword

ConfirmDeviceError
Errors returned by ConfirmDevice

ConfirmForgotPasswordError
Errors returned by ConfirmForgotPassword

ConfirmSignUpError
Errors returned by ConfirmSignUp

CreateGroupError
Errors returned by CreateGroup

CreateIdentityProviderError
Errors returned by CreateIdentityProvider

CreateResourceServerError
Errors returned by CreateResourceServer

CreateUserImportJobError
Errors returned by CreateUserImportJob

CreateUserPoolClientError
Errors returned by CreateUserPoolClient

CreateUserPoolDomainError
Errors returned by CreateUserPoolDomain

CreateUserPoolError
Errors returned by CreateUserPool

DeleteGroupError
Errors returned by DeleteGroup

DeleteIdentityProviderError
Errors returned by DeleteIdentityProvider

DeleteResourceServerError
Errors returned by DeleteResourceServer

DeleteUserAttributesError
Errors returned by DeleteUserAttributes

DeleteUserError
Errors returned by DeleteUser

DeleteUserPoolClientError
Errors returned by DeleteUserPoolClient

DeleteUserPoolDomainError
Errors returned by DeleteUserPoolDomain

DeleteUserPoolError
Errors returned by DeleteUserPool

DescribeIdentityProviderError
Errors returned by DescribeIdentityProvider

DescribeResourceServerError
Errors returned by DescribeResourceServer

DescribeRiskConfigurationError
Errors returned by DescribeRiskConfiguration

DescribeUserImportJobError
Errors returned by DescribeUserImportJob

DescribeUserPoolClientError
Errors returned by DescribeUserPoolClient

DescribeUserPoolDomainError
Errors returned by DescribeUserPoolDomain

DescribeUserPoolError
Errors returned by DescribeUserPool

ForgetDeviceError
Errors returned by ForgetDevice

ForgotPasswordError
Errors returned by ForgotPassword

GetCSVHeaderError
Errors returned by GetCSVHeader

GetDeviceError
Errors returned by GetDevice

GetGroupError
Errors returned by GetGroup

GetIdentityProviderByIdentifierError
Errors returned by GetIdentityProviderByIdentifier

GetSigningCertificateError
Errors returned by GetSigningCertificate

GetUICustomizationError
Errors returned by GetUICustomization

GetUserAttributeVerificationCodeError
Errors returned by GetUserAttributeVerificationCode

GetUserError
Errors returned by GetUser

GetUserPoolMfaConfigError
Errors returned by GetUserPoolMfaConfig

GlobalSignOutError
Errors returned by GlobalSignOut

InitiateAuthError
Errors returned by InitiateAuth

ListDevicesError
Errors returned by ListDevices

ListGroupsError
Errors returned by ListGroups

ListIdentityProvidersError
Errors returned by ListIdentityProviders

ListResourceServersError
Errors returned by ListResourceServers

ListUserImportJobsError
Errors returned by ListUserImportJobs

ListUserPoolClientsError
Errors returned by ListUserPoolClients

ListUserPoolsError
Errors returned by ListUserPools

ListUsersError
Errors returned by ListUsers

ListUsersInGroupError
Errors returned by ListUsersInGroup

ResendConfirmationCodeError
Errors returned by ResendConfirmationCode

RespondToAuthChallengeError
Errors returned by RespondToAuthChallenge

SetRiskConfigurationError
Errors returned by SetRiskConfiguration

SetUICustomizationError
Errors returned by SetUICustomization

SetUserMFAPreferenceError
Errors returned by SetUserMFAPreference

SetUserPoolMfaConfigError
Errors returned by SetUserPoolMfaConfig

SetUserSettingsError
Errors returned by SetUserSettings

SignUpError
Errors returned by SignUp

StartUserImportJobError
Errors returned by StartUserImportJob

StopUserImportJobError
Errors returned by StopUserImportJob

UpdateAuthEventFeedbackError
Errors returned by UpdateAuthEventFeedback

UpdateDeviceStatusError
Errors returned by UpdateDeviceStatus

UpdateGroupError
Errors returned by UpdateGroup

UpdateIdentityProviderError
Errors returned by UpdateIdentityProvider

UpdateResourceServerError
Errors returned by UpdateResourceServer

UpdateUserAttributesError
Errors returned by UpdateUserAttributes

UpdateUserPoolClientError
Errors returned by UpdateUserPoolClient

UpdateUserPoolError
Errors returned by UpdateUserPool

VerifySoftwareTokenError
Errors returned by VerifySoftwareToken

VerifyUserAttributeError
Errors returned by VerifyUserAttribute

Traits

CognitoIdentityProvider
Trait representing the capabilities of the Amazon Cognito Identity Provider API. Amazon Cognito Identity Provider clients implement this trait.