Struct rusoto_cognito_idp::CognitoIdentityProviderClient

pub struct CognitoIdentityProviderClient { /* fields omitted */ }

A client for the Amazon Cognito Identity Provider API.

Methods

impl CognitoIdentityProviderClient

pub fn new(region: Region) -> CognitoIdentityProviderClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CognitoIdentityProviderClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,