Struct rusoto_cognito_idp::CreateUserPoolClientRequest

pub struct CreateUserPoolClientRequest {
    pub allowed_o_auth_flows: Option<Vec<String>>,
    pub allowed_o_auth_flows_user_pool_client: Option<bool>,
    pub allowed_o_auth_scopes: Option<Vec<String>>,
    pub analytics_configuration: Option<AnalyticsConfigurationType>,
    pub callback_ur_ls: Option<Vec<String>>,
    pub client_name: String,
    pub default_redirect_uri: Option<String>,
    pub explicit_auth_flows: Option<Vec<String>>,
    pub generate_secret: Option<bool>,
    pub logout_ur_ls: Option<Vec<String>>,
    pub read_attributes: Option<Vec<String>>,
    pub refresh_token_validity: Option<i64>,
    pub supported_identity_providers: Option<Vec<String>>,
    pub user_pool_id: String,
    pub write_attributes: Option<Vec<String>>,
}

Represents the request to create a user pool client.

Fields
allowed_o_auth_flows: Option<Vec<String>>

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint.

Set to token to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

allowed_o_auth_flows_user_pool_client: Option<bool>

Set to True if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

allowed_o_auth_scopes: Option<Vec<String>>

A list of allowed OAuth scopes. Currently supported values are "phone", "email", "openid", and "Cognito".

analytics_configuration: Option<AnalyticsConfigurationType>

The Amazon Pinpoint analytics configuration for collecting metrics for this user pool.

callback_ur_ls: Option<Vec<String>>

A list of allowed redirect (callback) URLs for the identity providers.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

client_name: String

The client name for the user pool client you would like to create.

default_redirect_uri: Option<String>

The default redirect URI. Must be in the CallbackURLs list.

A redirect URI must:

Be an absolute URI.

Be registered with the authorization server.

Not include a fragment component.

See OAuth 2.0 - Redirection Endpoint.

Amazon Cognito requires HTTPS over HTTP except for http://localhost for testing purposes only.

App callback URLs such as myapp://example are also supported.

explicit_auth_flows: Option<Vec<String>>

The explicit authentication flows.

generate_secret: Option<bool>

Boolean to specify whether you want to generate a secret for the user pool client being created.

logout_ur_ls: Option<Vec<String>>

A list of allowed logout URLs for the identity providers.

read_attributes: Option<Vec<String>>

The read attributes.

refresh_token_validity: Option<i64>

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

supported_identity_providers: Option<Vec<String>>

A list of provider names for the identity providers that are supported on this client. The following are supported: COGNITO, Facebook, Google and LoginWithAmazon.

user_pool_id: String

The user pool ID for the user pool where you want to create a user pool client.

write_attributes: Option<Vec<String>>

The user pool attributes that the app client can write to.

If your app client allows users to sign in through an identity provider, this array must include all attributes that are mapped to identity provider attributes. Amazon Cognito updates mapped attributes when users sign in to your application through an identity provider. If your app client lacks write access to a mapped attribute, Amazon Cognito throws an error when it attempts to update the attribute. For more information, see Specifying Identity Provider Attribute Mappings for Your User Pool.