Trait rusoto_cognito_idp::CognitoIdentityProvider

pub trait CognitoIdentityProvider {
    fn add_custom_attributes(
        &self, 
        input: AddCustomAttributesRequest
    ) -> RusotoFuture<AddCustomAttributesResponse, AddCustomAttributesError>;

    fn admin_add_user_to_group(
        &self, 
        input: AdminAddUserToGroupRequest
    ) -> RusotoFuture<(), AdminAddUserToGroupError>;

    fn admin_confirm_sign_up(
        &self, 
        input: AdminConfirmSignUpRequest
    ) -> RusotoFuture<AdminConfirmSignUpResponse, AdminConfirmSignUpError>;

    fn admin_create_user(
        &self, 
        input: AdminCreateUserRequest
    ) -> RusotoFuture<AdminCreateUserResponse, AdminCreateUserError>;

    fn admin_delete_user(
        &self, 
        input: AdminDeleteUserRequest
    ) -> RusotoFuture<(), AdminDeleteUserError>;

    fn admin_delete_user_attributes(
        &self, 
        input: AdminDeleteUserAttributesRequest
    ) -> RusotoFuture<AdminDeleteUserAttributesResponse, AdminDeleteUserAttributesError>;

    fn admin_disable_provider_for_user(
        &self, 
        input: AdminDisableProviderForUserRequest
    ) -> RusotoFuture<AdminDisableProviderForUserResponse, AdminDisableProviderForUserError>;

    fn admin_disable_user(
        &self, 
        input: AdminDisableUserRequest
    ) -> RusotoFuture<AdminDisableUserResponse, AdminDisableUserError>;

    fn admin_enable_user(
        &self, 
        input: AdminEnableUserRequest
    ) -> RusotoFuture<AdminEnableUserResponse, AdminEnableUserError>;

    fn admin_forget_device(
        &self, 
        input: AdminForgetDeviceRequest
    ) -> RusotoFuture<(), AdminForgetDeviceError>;

    fn admin_get_device(
        &self, 
        input: AdminGetDeviceRequest
    ) -> RusotoFuture<AdminGetDeviceResponse, AdminGetDeviceError>;

    fn admin_get_user(
        &self, 
        input: AdminGetUserRequest
    ) -> RusotoFuture<AdminGetUserResponse, AdminGetUserError>;

    fn admin_initiate_auth(
        &self, 
        input: AdminInitiateAuthRequest
    ) -> RusotoFuture<AdminInitiateAuthResponse, AdminInitiateAuthError>;

    fn admin_link_provider_for_user(
        &self, 
        input: AdminLinkProviderForUserRequest
    ) -> RusotoFuture<AdminLinkProviderForUserResponse, AdminLinkProviderForUserError>;

    fn admin_list_devices(
        &self, 
        input: AdminListDevicesRequest
    ) -> RusotoFuture<AdminListDevicesResponse, AdminListDevicesError>;

    fn admin_list_groups_for_user(
        &self, 
        input: AdminListGroupsForUserRequest
    ) -> RusotoFuture<AdminListGroupsForUserResponse, AdminListGroupsForUserError>;

    fn admin_list_user_auth_events(
        &self, 
        input: AdminListUserAuthEventsRequest
    ) -> RusotoFuture<AdminListUserAuthEventsResponse, AdminListUserAuthEventsError>;

    fn admin_remove_user_from_group(
        &self, 
        input: AdminRemoveUserFromGroupRequest
    ) -> RusotoFuture<(), AdminRemoveUserFromGroupError>;

    fn admin_reset_user_password(
        &self, 
        input: AdminResetUserPasswordRequest
    ) -> RusotoFuture<AdminResetUserPasswordResponse, AdminResetUserPasswordError>;

    fn admin_respond_to_auth_challenge(
        &self, 
        input: AdminRespondToAuthChallengeRequest
    ) -> RusotoFuture<AdminRespondToAuthChallengeResponse, AdminRespondToAuthChallengeError>;

    fn admin_set_user_mfa_preference(
        &self, 
        input: AdminSetUserMFAPreferenceRequest
    ) -> RusotoFuture<AdminSetUserMFAPreferenceResponse, AdminSetUserMFAPreferenceError>;

    fn admin_set_user_settings(
        &self, 
        input: AdminSetUserSettingsRequest
    ) -> RusotoFuture<AdminSetUserSettingsResponse, AdminSetUserSettingsError>;

    fn admin_update_auth_event_feedback(
        &self, 
        input: AdminUpdateAuthEventFeedbackRequest
    ) -> RusotoFuture<AdminUpdateAuthEventFeedbackResponse, AdminUpdateAuthEventFeedbackError>;

    fn admin_update_device_status(
        &self, 
        input: AdminUpdateDeviceStatusRequest
    ) -> RusotoFuture<AdminUpdateDeviceStatusResponse, AdminUpdateDeviceStatusError>;

    fn admin_update_user_attributes(
        &self, 
        input: AdminUpdateUserAttributesRequest
    ) -> RusotoFuture<AdminUpdateUserAttributesResponse, AdminUpdateUserAttributesError>;

    fn admin_user_global_sign_out(
        &self, 
        input: AdminUserGlobalSignOutRequest
    ) -> RusotoFuture<AdminUserGlobalSignOutResponse, AdminUserGlobalSignOutError>;

    fn associate_software_token(
        &self, 
        input: AssociateSoftwareTokenRequest
    ) -> RusotoFuture<AssociateSoftwareTokenResponse, AssociateSoftwareTokenError>;

    fn change_password(
        &self, 
        input: ChangePasswordRequest
    ) -> RusotoFuture<ChangePasswordResponse, ChangePasswordError>;

    fn confirm_device(
        &self, 
        input: ConfirmDeviceRequest
    ) -> RusotoFuture<ConfirmDeviceResponse, ConfirmDeviceError>;

    fn confirm_forgot_password(
        &self, 
        input: ConfirmForgotPasswordRequest
    ) -> RusotoFuture<ConfirmForgotPasswordResponse, ConfirmForgotPasswordError>;

    fn confirm_sign_up(
        &self, 
        input: ConfirmSignUpRequest
    ) -> RusotoFuture<ConfirmSignUpResponse, ConfirmSignUpError>;

    fn create_group(
        &self, 
        input: CreateGroupRequest
    ) -> RusotoFuture<CreateGroupResponse, CreateGroupError>;

    fn create_identity_provider(
        &self, 
        input: CreateIdentityProviderRequest
    ) -> RusotoFuture<CreateIdentityProviderResponse, CreateIdentityProviderError>;

    fn create_resource_server(
        &self, 
        input: CreateResourceServerRequest
    ) -> RusotoFuture<CreateResourceServerResponse, CreateResourceServerError>;

    fn create_user_import_job(
        &self, 
        input: CreateUserImportJobRequest
    ) -> RusotoFuture<CreateUserImportJobResponse, CreateUserImportJobError>;

    fn create_user_pool(
        &self, 
        input: CreateUserPoolRequest
    ) -> RusotoFuture<CreateUserPoolResponse, CreateUserPoolError>;

    fn create_user_pool_client(
        &self, 
        input: CreateUserPoolClientRequest
    ) -> RusotoFuture<CreateUserPoolClientResponse, CreateUserPoolClientError>;

    fn create_user_pool_domain(
        &self, 
        input: CreateUserPoolDomainRequest
    ) -> RusotoFuture<CreateUserPoolDomainResponse, CreateUserPoolDomainError>;

    fn delete_group(
        &self, 
        input: DeleteGroupRequest
    ) -> RusotoFuture<(), DeleteGroupError>;

    fn delete_identity_provider(
        &self, 
        input: DeleteIdentityProviderRequest
    ) -> RusotoFuture<(), DeleteIdentityProviderError>;

    fn delete_resource_server(
        &self, 
        input: DeleteResourceServerRequest
    ) -> RusotoFuture<(), DeleteResourceServerError>;

    fn delete_user(
        &self, 
        input: DeleteUserRequest
    ) -> RusotoFuture<(), DeleteUserError>;

    fn delete_user_attributes(
        &self, 
        input: DeleteUserAttributesRequest
    ) -> RusotoFuture<DeleteUserAttributesResponse, DeleteUserAttributesError>;

    fn delete_user_pool(
        &self, 
        input: DeleteUserPoolRequest
    ) -> RusotoFuture<(), DeleteUserPoolError>;

    fn delete_user_pool_client(
        &self, 
        input: DeleteUserPoolClientRequest
    ) -> RusotoFuture<(), DeleteUserPoolClientError>;

    fn delete_user_pool_domain(
        &self, 
        input: DeleteUserPoolDomainRequest
    ) -> RusotoFuture<DeleteUserPoolDomainResponse, DeleteUserPoolDomainError>;

    fn describe_identity_provider(
        &self, 
        input: DescribeIdentityProviderRequest
    ) -> RusotoFuture<DescribeIdentityProviderResponse, DescribeIdentityProviderError>;

    fn describe_resource_server(
        &self, 
        input: DescribeResourceServerRequest
    ) -> RusotoFuture<DescribeResourceServerResponse, DescribeResourceServerError>;

    fn describe_risk_configuration(
        &self, 
        input: DescribeRiskConfigurationRequest
    ) -> RusotoFuture<DescribeRiskConfigurationResponse, DescribeRiskConfigurationError>;

    fn describe_user_import_job(
        &self, 
        input: DescribeUserImportJobRequest
    ) -> RusotoFuture<DescribeUserImportJobResponse, DescribeUserImportJobError>;

    fn describe_user_pool(
        &self, 
        input: DescribeUserPoolRequest
    ) -> RusotoFuture<DescribeUserPoolResponse, DescribeUserPoolError>;

    fn describe_user_pool_client(
        &self, 
        input: DescribeUserPoolClientRequest
    ) -> RusotoFuture<DescribeUserPoolClientResponse, DescribeUserPoolClientError>;

    fn describe_user_pool_domain(
        &self, 
        input: DescribeUserPoolDomainRequest
    ) -> RusotoFuture<DescribeUserPoolDomainResponse, DescribeUserPoolDomainError>;

    fn forget_device(
        &self, 
        input: ForgetDeviceRequest
    ) -> RusotoFuture<(), ForgetDeviceError>;

    fn forgot_password(
        &self, 
        input: ForgotPasswordRequest
    ) -> RusotoFuture<ForgotPasswordResponse, ForgotPasswordError>;

    fn get_csv_header(
        &self, 
        input: GetCSVHeaderRequest
    ) -> RusotoFuture<GetCSVHeaderResponse, GetCSVHeaderError>;

    fn get_device(
        &self, 
        input: GetDeviceRequest
    ) -> RusotoFuture<GetDeviceResponse, GetDeviceError>;

    fn get_group(
        &self, 
        input: GetGroupRequest
    ) -> RusotoFuture<GetGroupResponse, GetGroupError>;

    fn get_identity_provider_by_identifier(
        &self, 
        input: GetIdentityProviderByIdentifierRequest
    ) -> RusotoFuture<GetIdentityProviderByIdentifierResponse, GetIdentityProviderByIdentifierError>;

    fn get_signing_certificate(
        &self, 
        input: GetSigningCertificateRequest
    ) -> RusotoFuture<GetSigningCertificateResponse, GetSigningCertificateError>;

    fn get_ui_customization(
        &self, 
        input: GetUICustomizationRequest
    ) -> RusotoFuture<GetUICustomizationResponse, GetUICustomizationError>;

    fn get_user(
        &self, 
        input: GetUserRequest
    ) -> RusotoFuture<GetUserResponse, GetUserError>;

    fn get_user_attribute_verification_code(
        &self, 
        input: GetUserAttributeVerificationCodeRequest
    ) -> RusotoFuture<GetUserAttributeVerificationCodeResponse, GetUserAttributeVerificationCodeError>;

    fn get_user_pool_mfa_config(
        &self, 
        input: GetUserPoolMfaConfigRequest
    ) -> RusotoFuture<GetUserPoolMfaConfigResponse, GetUserPoolMfaConfigError>;

    fn global_sign_out(
        &self, 
        input: GlobalSignOutRequest
    ) -> RusotoFuture<GlobalSignOutResponse, GlobalSignOutError>;

    fn initiate_auth(
        &self, 
        input: InitiateAuthRequest
    ) -> RusotoFuture<InitiateAuthResponse, InitiateAuthError>;

    fn list_devices(
        &self, 
        input: ListDevicesRequest
    ) -> RusotoFuture<ListDevicesResponse, ListDevicesError>;

    fn list_groups(
        &self, 
        input: ListGroupsRequest
    ) -> RusotoFuture<ListGroupsResponse, ListGroupsError>;

    fn list_identity_providers(
        &self, 
        input: ListIdentityProvidersRequest
    ) -> RusotoFuture<ListIdentityProvidersResponse, ListIdentityProvidersError>;

    fn list_resource_servers(
        &self, 
        input: ListResourceServersRequest
    ) -> RusotoFuture<ListResourceServersResponse, ListResourceServersError>;

    fn list_user_import_jobs(
        &self, 
        input: ListUserImportJobsRequest
    ) -> RusotoFuture<ListUserImportJobsResponse, ListUserImportJobsError>;

    fn list_user_pool_clients(
        &self, 
        input: ListUserPoolClientsRequest
    ) -> RusotoFuture<ListUserPoolClientsResponse, ListUserPoolClientsError>;

    fn list_user_pools(
        &self, 
        input: ListUserPoolsRequest
    ) -> RusotoFuture<ListUserPoolsResponse, ListUserPoolsError>;

    fn list_users(
        &self, 
        input: ListUsersRequest
    ) -> RusotoFuture<ListUsersResponse, ListUsersError>;

    fn list_users_in_group(
        &self, 
        input: ListUsersInGroupRequest
    ) -> RusotoFuture<ListUsersInGroupResponse, ListUsersInGroupError>;

    fn resend_confirmation_code(
        &self, 
        input: ResendConfirmationCodeRequest
    ) -> RusotoFuture<ResendConfirmationCodeResponse, ResendConfirmationCodeError>;

    fn respond_to_auth_challenge(
        &self, 
        input: RespondToAuthChallengeRequest
    ) -> RusotoFuture<RespondToAuthChallengeResponse, RespondToAuthChallengeError>;

    fn set_risk_configuration(
        &self, 
        input: SetRiskConfigurationRequest
    ) -> RusotoFuture<SetRiskConfigurationResponse, SetRiskConfigurationError>;

    fn set_ui_customization(
        &self, 
        input: SetUICustomizationRequest
    ) -> RusotoFuture<SetUICustomizationResponse, SetUICustomizationError>;

    fn set_user_mfa_preference(
        &self, 
        input: SetUserMFAPreferenceRequest
    ) -> RusotoFuture<SetUserMFAPreferenceResponse, SetUserMFAPreferenceError>;

    fn set_user_pool_mfa_config(
        &self, 
        input: SetUserPoolMfaConfigRequest
    ) -> RusotoFuture<SetUserPoolMfaConfigResponse, SetUserPoolMfaConfigError>;

    fn set_user_settings(
        &self, 
        input: SetUserSettingsRequest
    ) -> RusotoFuture<SetUserSettingsResponse, SetUserSettingsError>;

    fn sign_up(
        &self, 
        input: SignUpRequest
    ) -> RusotoFuture<SignUpResponse, SignUpError>;

    fn start_user_import_job(
        &self, 
        input: StartUserImportJobRequest
    ) -> RusotoFuture<StartUserImportJobResponse, StartUserImportJobError>;

    fn stop_user_import_job(
        &self, 
        input: StopUserImportJobRequest
    ) -> RusotoFuture<StopUserImportJobResponse, StopUserImportJobError>;

    fn update_auth_event_feedback(
        &self, 
        input: UpdateAuthEventFeedbackRequest
    ) -> RusotoFuture<UpdateAuthEventFeedbackResponse, UpdateAuthEventFeedbackError>;

    fn update_device_status(
        &self, 
        input: UpdateDeviceStatusRequest
    ) -> RusotoFuture<UpdateDeviceStatusResponse, UpdateDeviceStatusError>;

    fn update_group(
        &self, 
        input: UpdateGroupRequest
    ) -> RusotoFuture<UpdateGroupResponse, UpdateGroupError>;

    fn update_identity_provider(
        &self, 
        input: UpdateIdentityProviderRequest
    ) -> RusotoFuture<UpdateIdentityProviderResponse, UpdateIdentityProviderError>;

    fn update_resource_server(
        &self, 
        input: UpdateResourceServerRequest
    ) -> RusotoFuture<UpdateResourceServerResponse, UpdateResourceServerError>;

    fn update_user_attributes(
        &self, 
        input: UpdateUserAttributesRequest
    ) -> RusotoFuture<UpdateUserAttributesResponse, UpdateUserAttributesError>;

    fn update_user_pool(
        &self, 
        input: UpdateUserPoolRequest
    ) -> RusotoFuture<UpdateUserPoolResponse, UpdateUserPoolError>;

    fn update_user_pool_client(
        &self, 
        input: UpdateUserPoolClientRequest
    ) -> RusotoFuture<UpdateUserPoolClientResponse, UpdateUserPoolClientError>;

    fn verify_software_token(
        &self, 
        input: VerifySoftwareTokenRequest
    ) -> RusotoFuture<VerifySoftwareTokenResponse, VerifySoftwareTokenError>;

    fn verify_user_attribute(
        &self, 
        input: VerifyUserAttributeRequest
    ) -> RusotoFuture<VerifyUserAttributeResponse, VerifyUserAttributeError>;
}

Trait representing the capabilities of the Amazon Cognito Identity Provider API. Amazon Cognito Identity Provider clients implement this trait.

Required Methods

fn add_custom_attributes(
&self,
input: AddCustomAttributesRequest
) -> RusotoFuture<AddCustomAttributesResponse, AddCustomAttributesError>

Adds additional user attributes to the user pool schema.

fn admin_add_user_to_group(
&self,
input: AdminAddUserToGroupRequest
) -> RusotoFuture<(), AdminAddUserToGroupError>

Adds the specified user to the specified group.

Requires developer credentials.

fn admin_confirm_sign_up(
&self,
input: AdminConfirmSignUpRequest
) -> RusotoFuture<AdminConfirmSignUpResponse, AdminConfirmSignUpError>

Confirms user registration as an admin without using a confirmation code. Works on any user.

Requires developer credentials.

fn admin_create_user(
&self,
input: AdminCreateUserRequest
) -> RusotoFuture<AdminCreateUserResponse, AdminCreateUserError>

Creates a new user in the specified user pool.

If MessageAction is not set, the default is to send a welcome message via email or phone (SMS).

This message is based on a template that you configured in your call to or . This template includes your custom sign-up instructions and placeholders for user name and temporary password.

Alternatively, you can call AdminCreateUser with “SUPPRESS” for the MessageAction parameter, and Amazon Cognito will not send any email.

In either case, the user will be in the FORCE_CHANGE_PASSWORD state until they sign in and change their password.

AdminCreateUser requires developer credentials.

fn admin_delete_user(
&self,
input: AdminDeleteUserRequest
) -> RusotoFuture<(), AdminDeleteUserError>

Deletes a user as an administrator. Works on any user.

Requires developer credentials.

fn admin_delete_user_attributes(
&self,
input: AdminDeleteUserAttributesRequest
) -> RusotoFuture<AdminDeleteUserAttributesResponse, AdminDeleteUserAttributesError>

Deletes the user attributes in a user pool as an administrator. Works on any user.

Requires developer credentials.

fn admin_disable_provider_for_user(
&self,
input: AdminDisableProviderForUserRequest
) -> RusotoFuture<AdminDisableProviderForUserResponse, AdminDisableProviderForUserError>

Disables the user from signing in with the specified external (SAML or social) identity provider. If the user to disable is a Cognito User Pools native username + password user, they are not permitted to use their password to sign-in. If the user to disable is a linked external IdP user, any link between that user and an existing user is removed. The next time the external user (no longer attached to the previously linked DestinationUser) signs in, they must create a new user account. See .

This action is enabled only for admin access and requires developer credentials.

The ProviderName must match the value specified when creating an IdP for the pool.

To disable a native username + password user, the ProviderName value must be Cognito and the ProviderAttributeName must be Cognito_Subject, with the ProviderAttributeValue being the name that is used in the user pool for the user.

The ProviderAttributeName must always be Cognito_Subject for social identity providers. The ProviderAttributeValue must always be the exact subject that was used when the user was originally linked as a source user.

For de-linking a SAML identity, there are two scenarios. If the linked identity has not yet been used to sign-in, the ProviderAttributeName and ProviderAttributeValue must be the same values that were used for the SourceUser when the identities were originally linked in the call. (If the linking was done with ProviderAttributeName set to Cognito_Subject, the same applies here). However, if the user has already signed in, the ProviderAttributeName must be Cognito_Subject and ProviderAttributeValue must be the subject of the SAML assertion.

fn admin_disable_user(
&self,
input: AdminDisableUserRequest
) -> RusotoFuture<AdminDisableUserResponse, AdminDisableUserError>

Disables the specified user as an administrator. Works on any user.

Requires developer credentials.

fn admin_enable_user(
&self,
input: AdminEnableUserRequest
) -> RusotoFuture<AdminEnableUserResponse, AdminEnableUserError>

Enables the specified user as an administrator. Works on any user.

Requires developer credentials.

fn admin_forget_device(
&self,
input: AdminForgetDeviceRequest
) -> RusotoFuture<(), AdminForgetDeviceError>

Forgets the device, as an administrator.

Requires developer credentials.

fn admin_get_device(
&self,
input: AdminGetDeviceRequest
) -> RusotoFuture<AdminGetDeviceResponse, AdminGetDeviceError>

Gets the device, as an administrator.

Requires developer credentials.

fn admin_get_user(
&self,
input: AdminGetUserRequest
) -> RusotoFuture<AdminGetUserResponse, AdminGetUserError>

Gets the specified user by user name in a user pool as an administrator. Works on any user.

Requires developer credentials.

fn admin_initiate_auth(
&self,
input: AdminInitiateAuthRequest
) -> RusotoFuture<AdminInitiateAuthResponse, AdminInitiateAuthError>

Initiates the authentication flow, as an administrator.

Requires developer credentials.

fn admin_link_provider_for_user(
&self,
input: AdminLinkProviderForUserRequest
) -> RusotoFuture<AdminLinkProviderForUserResponse, AdminLinkProviderForUserError>

Links an existing user account in a user pool (DestinationUser) to an identity from an external identity provider (SourceUser) based on a specified attribute name and value from the external identity provider. This allows you to create a link from the existing user account to an external federated user identity that has not yet been used to sign in, so that the federated user identity can be used to sign in as the existing user account.

For example, if there is an existing user with a username and password, this API links that user to a federated user identity, so that when the federated user identity is used, the user signs in as the existing user account.

Because this API allows a user with an external federated identity to sign in as an existing user in the user pool, it is critical that it only be used with external identity providers and provider attributes that have been trusted by the application owner.

See also .

This action is enabled only for admin access and requires developer credentials.

fn admin_list_devices(
&self,
input: AdminListDevicesRequest
) -> RusotoFuture<AdminListDevicesResponse, AdminListDevicesError>

Lists devices, as an administrator.

Requires developer credentials.

fn admin_list_groups_for_user(
&self,
input: AdminListGroupsForUserRequest
) -> RusotoFuture<AdminListGroupsForUserResponse, AdminListGroupsForUserError>

Lists the groups that the user belongs to.

Requires developer credentials.

fn admin_list_user_auth_events(
&self,
input: AdminListUserAuthEventsRequest
) -> RusotoFuture<AdminListUserAuthEventsResponse, AdminListUserAuthEventsError>

Lists a history of user activity and any risks detected as part of Amazon Cognito advanced security.

fn admin_remove_user_from_group(
&self,
input: AdminRemoveUserFromGroupRequest
) -> RusotoFuture<(), AdminRemoveUserFromGroupError>

Removes the specified user from the specified group.

Requires developer credentials.

fn admin_reset_user_password(
&self,
input: AdminResetUserPasswordRequest
) -> RusotoFuture<AdminResetUserPasswordResponse, AdminResetUserPasswordError>

Resets the specified user's password in a user pool as an administrator. Works on any user.

When a developer calls this API, the current password is invalidated, so it must be changed. If a user tries to sign in after the API is called, the app will get a PasswordResetRequiredException exception back and should direct the user down the flow to reset the password, which is the same as the forgot password flow. In addition, if the user pool has phone verification selected and a verified phone number exists for the user, or if email verification is selected and a verified email exists for the user, calling this API will also result in sending a message to the end user with the code to change their password.

Requires developer credentials.

fn admin_respond_to_auth_challenge(
&self,
input: AdminRespondToAuthChallengeRequest
) -> RusotoFuture<AdminRespondToAuthChallengeResponse, AdminRespondToAuthChallengeError>

Responds to an authentication challenge, as an administrator.

Requires developer credentials.

fn admin_set_user_mfa_preference(
&self,
input: AdminSetUserMFAPreferenceRequest
) -> RusotoFuture<AdminSetUserMFAPreferenceResponse, AdminSetUserMFAPreferenceError>

Sets the user's multi-factor authentication (MFA) preference.

fn admin_set_user_settings(
&self,
input: AdminSetUserSettingsRequest
) -> RusotoFuture<AdminSetUserSettingsResponse, AdminSetUserSettingsError>

Sets all the user settings for a specified user name. Works on any user.

Requires developer credentials.

fn admin_update_auth_event_feedback(
&self,
input: AdminUpdateAuthEventFeedbackRequest
) -> RusotoFuture<AdminUpdateAuthEventFeedbackResponse, AdminUpdateAuthEventFeedbackError>

Provides feedback for an authentication event as to whether it was from a valid user. This feedback is used for improving the risk evaluation decision for the user pool as part of Amazon Cognito advanced security.

fn admin_update_device_status(
&self,
input: AdminUpdateDeviceStatusRequest
) -> RusotoFuture<AdminUpdateDeviceStatusResponse, AdminUpdateDeviceStatusError>

Updates the device status as an administrator.

Requires developer credentials.

fn admin_update_user_attributes(
&self,
input: AdminUpdateUserAttributesRequest
) -> RusotoFuture<AdminUpdateUserAttributesResponse, AdminUpdateUserAttributesError>

Updates the specified user's attributes, including developer attributes, as an administrator. Works on any user.

For custom attributes, you must prepend the custom: prefix to the attribute name.

In addition to updating user attributes, this API can also be used to mark phone and email as verified.

Requires developer credentials.

fn admin_user_global_sign_out(
&self,
input: AdminUserGlobalSignOutRequest
) -> RusotoFuture<AdminUserGlobalSignOutResponse, AdminUserGlobalSignOutError>

Signs out users from all devices, as an administrator.

Requires developer credentials.

fn associate_software_token(
&self,
input: AssociateSoftwareTokenRequest
) -> RusotoFuture<AssociateSoftwareTokenResponse, AssociateSoftwareTokenError>

Returns a unique generated shared secret key code for the user account. The request takes an access token or a session string, but not both.

fn change_password(
&self,
input: ChangePasswordRequest
) -> RusotoFuture<ChangePasswordResponse, ChangePasswordError>

Changes the password for a specified user in a user pool.

fn confirm_device(
&self,
input: ConfirmDeviceRequest
) -> RusotoFuture<ConfirmDeviceResponse, ConfirmDeviceError>

Confirms tracking of the device. This API call is the call that begins device tracking.

fn confirm_forgot_password(
&self,
input: ConfirmForgotPasswordRequest
) -> RusotoFuture<ConfirmForgotPasswordResponse, ConfirmForgotPasswordError>

Allows a user to enter a confirmation code to reset a forgotten password.

fn confirm_sign_up(
&self,
input: ConfirmSignUpRequest
) -> RusotoFuture<ConfirmSignUpResponse, ConfirmSignUpError>

Confirms registration of a user and handles the existing alias from a previous user.

fn create_group(
&self,
input: CreateGroupRequest
) -> RusotoFuture<CreateGroupResponse, CreateGroupError>

Creates a new group in the specified user pool.

Requires developer credentials.

fn create_identity_provider(
&self,
input: CreateIdentityProviderRequest
) -> RusotoFuture<CreateIdentityProviderResponse, CreateIdentityProviderError>

Creates an identity provider for a user pool.

fn create_resource_server(
&self,
input: CreateResourceServerRequest
) -> RusotoFuture<CreateResourceServerResponse, CreateResourceServerError>

Creates a new OAuth2.0 resource server and defines custom scopes in it.

fn create_user_import_job(
&self,
input: CreateUserImportJobRequest
) -> RusotoFuture<CreateUserImportJobResponse, CreateUserImportJobError>

Creates the user import job.

fn create_user_pool(
&self,
input: CreateUserPoolRequest
) -> RusotoFuture<CreateUserPoolResponse, CreateUserPoolError>

Creates a new Amazon Cognito user pool and sets the password policy for the pool.

fn create_user_pool_client(
&self,
input: CreateUserPoolClientRequest
) -> RusotoFuture<CreateUserPoolClientResponse, CreateUserPoolClientError>

Creates the user pool client.

fn create_user_pool_domain(
&self,
input: CreateUserPoolDomainRequest
) -> RusotoFuture<CreateUserPoolDomainResponse, CreateUserPoolDomainError>

Creates a new domain for a user pool.

fn delete_group(
&self,
input: DeleteGroupRequest
) -> RusotoFuture<(), DeleteGroupError>

Deletes a group. Currently only groups with no members can be deleted.

Requires developer credentials.

fn delete_identity_provider(
&self,
input: DeleteIdentityProviderRequest
) -> RusotoFuture<(), DeleteIdentityProviderError>

Deletes an identity provider for a user pool.

fn delete_resource_server(
&self,
input: DeleteResourceServerRequest
) -> RusotoFuture<(), DeleteResourceServerError>

Deletes a resource server.

fn delete_user(
&self,
input: DeleteUserRequest
) -> RusotoFuture<(), DeleteUserError>

Allows a user to delete himself or herself.

fn delete_user_attributes(
&self,
input: DeleteUserAttributesRequest
) -> RusotoFuture<DeleteUserAttributesResponse, DeleteUserAttributesError>

Deletes the attributes for a user.

fn delete_user_pool(
&self,
input: DeleteUserPoolRequest
) -> RusotoFuture<(), DeleteUserPoolError>

Deletes the specified Amazon Cognito user pool.

fn delete_user_pool_client(
&self,
input: DeleteUserPoolClientRequest
) -> RusotoFuture<(), DeleteUserPoolClientError>

Allows the developer to delete the user pool client.

fn delete_user_pool_domain(
&self,
input: DeleteUserPoolDomainRequest
) -> RusotoFuture<DeleteUserPoolDomainResponse, DeleteUserPoolDomainError>

Deletes a domain for a user pool.

fn describe_identity_provider(
&self,
input: DescribeIdentityProviderRequest
) -> RusotoFuture<DescribeIdentityProviderResponse, DescribeIdentityProviderError>

Gets information about a specific identity provider.

fn describe_resource_server(
&self,
input: DescribeResourceServerRequest
) -> RusotoFuture<DescribeResourceServerResponse, DescribeResourceServerError>

Describes a resource server.

fn describe_risk_configuration(
&self,
input: DescribeRiskConfigurationRequest
) -> RusotoFuture<DescribeRiskConfigurationResponse, DescribeRiskConfigurationError>

Describes the risk configuration.

fn describe_user_import_job(
&self,
input: DescribeUserImportJobRequest
) -> RusotoFuture<DescribeUserImportJobResponse, DescribeUserImportJobError>

Describes the user import job.

fn describe_user_pool(
&self,
input: DescribeUserPoolRequest
) -> RusotoFuture<DescribeUserPoolResponse, DescribeUserPoolError>

Returns the configuration information and metadata of the specified user pool.

fn describe_user_pool_client(
&self,
input: DescribeUserPoolClientRequest
) -> RusotoFuture<DescribeUserPoolClientResponse, DescribeUserPoolClientError>

Client method for returning the configuration information and metadata of the specified user pool client.

fn describe_user_pool_domain(
&self,
input: DescribeUserPoolDomainRequest
) -> RusotoFuture<DescribeUserPoolDomainResponse, DescribeUserPoolDomainError>

Gets information about a domain.

fn forget_device(
&self,
input: ForgetDeviceRequest
) -> RusotoFuture<(), ForgetDeviceError>

Forgets the specified device.

fn forgot_password(
&self,
input: ForgotPasswordRequest
) -> RusotoFuture<ForgotPasswordResponse, ForgotPasswordError>

Calling this API causes a message to be sent to the end user with a confirmation code that is required to change the user's password. For the Username parameter, you can use the username or user alias. If a verified phone number exists for the user, the confirmation code is sent to the phone number. Otherwise, if a verified email exists, the confirmation code is sent to the email. If neither a verified phone number nor a verified email exists, InvalidParameterException is thrown. To use the confirmation code for resetting the password, call .

fn get_csv_header(
&self,
input: GetCSVHeaderRequest
) -> RusotoFuture<GetCSVHeaderResponse, GetCSVHeaderError>

Gets the header information for the .csv file to be used as input for the user import job.

fn get_device(
&self,
input: GetDeviceRequest
) -> RusotoFuture<GetDeviceResponse, GetDeviceError>

Gets the device.

fn get_group(
&self,
input: GetGroupRequest
) -> RusotoFuture<GetGroupResponse, GetGroupError>

Gets a group.

Requires developer credentials.

fn get_identity_provider_by_identifier(
&self,
input: GetIdentityProviderByIdentifierRequest
) -> RusotoFuture<GetIdentityProviderByIdentifierResponse, GetIdentityProviderByIdentifierError>

Gets the specified identity provider.

fn get_signing_certificate(
&self,
input: GetSigningCertificateRequest
) -> RusotoFuture<GetSigningCertificateResponse, GetSigningCertificateError>

This method takes a user pool ID, and returns the signing certificate.

fn get_ui_customization(
&self,
input: GetUICustomizationRequest
) -> RusotoFuture<GetUICustomizationResponse, GetUICustomizationError>

Gets the UI Customization information for a particular app client's app UI, if there is something set. If nothing is set for the particular client, but there is an existing pool level customization (app clientId will be ALL), then that is returned. If nothing is present, then an empty shape is returned.

fn get_user(
&self,
input: GetUserRequest
) -> RusotoFuture<GetUserResponse, GetUserError>

Gets the user attributes and metadata for a user.

fn get_user_attribute_verification_code(
&self,
input: GetUserAttributeVerificationCodeRequest
) -> RusotoFuture<GetUserAttributeVerificationCodeResponse, GetUserAttributeVerificationCodeError>

Gets the user attribute verification code for the specified attribute name.

fn get_user_pool_mfa_config(
&self,
input: GetUserPoolMfaConfigRequest
) -> RusotoFuture<GetUserPoolMfaConfigResponse, GetUserPoolMfaConfigError>

Gets the user pool multi-factor authentication (MFA) configuration.

fn global_sign_out(
&self,
input: GlobalSignOutRequest
) -> RusotoFuture<GlobalSignOutResponse, GlobalSignOutError>

Signs out users from all devices.

fn initiate_auth(
&self,
input: InitiateAuthRequest
) -> RusotoFuture<InitiateAuthResponse, InitiateAuthError>

Initiates the authentication flow.

fn list_devices(
&self,
input: ListDevicesRequest
) -> RusotoFuture<ListDevicesResponse, ListDevicesError>

Lists the devices.

fn list_groups(
&self,
input: ListGroupsRequest
) -> RusotoFuture<ListGroupsResponse, ListGroupsError>

Lists the groups associated with a user pool.

Requires developer credentials.

fn list_identity_providers(
&self,
input: ListIdentityProvidersRequest
) -> RusotoFuture<ListIdentityProvidersResponse, ListIdentityProvidersError>

Lists information about all identity providers for a user pool.

fn list_resource_servers(
&self,
input: ListResourceServersRequest
) -> RusotoFuture<ListResourceServersResponse, ListResourceServersError>

Lists the resource servers for a user pool.

fn list_user_import_jobs(
&self,
input: ListUserImportJobsRequest
) -> RusotoFuture<ListUserImportJobsResponse, ListUserImportJobsError>

Lists the user import jobs.

fn list_user_pool_clients(
&self,
input: ListUserPoolClientsRequest
) -> RusotoFuture<ListUserPoolClientsResponse, ListUserPoolClientsError>

Lists the clients that have been created for the specified user pool.

fn list_user_pools(
&self,
input: ListUserPoolsRequest
) -> RusotoFuture<ListUserPoolsResponse, ListUserPoolsError>

Lists the user pools associated with an AWS account.

fn list_users(
&self,
input: ListUsersRequest
) -> RusotoFuture<ListUsersResponse, ListUsersError>

Lists the users in the Amazon Cognito user pool.

fn list_users_in_group(
&self,
input: ListUsersInGroupRequest
) -> RusotoFuture<ListUsersInGroupResponse, ListUsersInGroupError>

Lists the users in the specified group.

Requires developer credentials.

fn resend_confirmation_code(
&self,
input: ResendConfirmationCodeRequest
) -> RusotoFuture<ResendConfirmationCodeResponse, ResendConfirmationCodeError>

Resends the confirmation (for confirmation of registration) to a specific user in the user pool.

fn respond_to_auth_challenge(
&self,
input: RespondToAuthChallengeRequest
) -> RusotoFuture<RespondToAuthChallengeResponse, RespondToAuthChallengeError>

Responds to the authentication challenge.

fn set_risk_configuration(
&self,
input: SetRiskConfigurationRequest
) -> RusotoFuture<SetRiskConfigurationResponse, SetRiskConfigurationError>

Configures actions on detected risks. To delete the risk configuration for UserPoolId or ClientId, pass null values for all four configuration types.

To enable Amazon Cognito advanced security features, update the user pool to include the UserPoolAddOns keyAdvancedSecurityMode.

See .

fn set_ui_customization(
&self,
input: SetUICustomizationRequest
) -> RusotoFuture<SetUICustomizationResponse, SetUICustomizationError>

Sets the UI customization information for a user pool's built-in app UI.

You can specify app UI customization settings for a single client (with a specific clientId) or for all clients (by setting the clientId to ALL). If you specify ALL, the default configuration will be used for every client that has no UI customization set previously. If you specify UI customization settings for a particular client, it will no longer fall back to the ALL configuration.

To use this API, your user pool must have a domain associated with it. Otherwise, there is no place to host the app's pages, and the service will throw an error.

fn set_user_mfa_preference(
&self,
input: SetUserMFAPreferenceRequest
) -> RusotoFuture<SetUserMFAPreferenceResponse, SetUserMFAPreferenceError>

Set the user's multi-factor authentication (MFA) method preference.

fn set_user_pool_mfa_config(
&self,
input: SetUserPoolMfaConfigRequest
) -> RusotoFuture<SetUserPoolMfaConfigResponse, SetUserPoolMfaConfigError>

Set the user pool MFA configuration.

fn set_user_settings(
&self,
input: SetUserSettingsRequest
) -> RusotoFuture<SetUserSettingsResponse, SetUserSettingsError>

Sets the user settings like multi-factor authentication (MFA). If MFA is to be removed for a particular attribute pass the attribute with code delivery as null. If null list is passed, all MFA options are removed.

fn sign_up(
&self,
input: SignUpRequest
) -> RusotoFuture<SignUpResponse, SignUpError>

Registers the user in the specified user pool and creates a user name, password, and user attributes.

fn start_user_import_job(
&self,
input: StartUserImportJobRequest
) -> RusotoFuture<StartUserImportJobResponse, StartUserImportJobError>

Starts the user import.

fn stop_user_import_job(
&self,
input: StopUserImportJobRequest
) -> RusotoFuture<StopUserImportJobResponse, StopUserImportJobError>

Stops the user import job.

fn update_auth_event_feedback(
&self,
input: UpdateAuthEventFeedbackRequest
) -> RusotoFuture<UpdateAuthEventFeedbackResponse, UpdateAuthEventFeedbackError>

Provides the feedback for an authentication event whether it was from a valid user or not. This feedback is used for improving the risk evaluation decision for the user pool as part of Amazon Cognito advanced security.

fn update_device_status(
&self,
input: UpdateDeviceStatusRequest
) -> RusotoFuture<UpdateDeviceStatusResponse, UpdateDeviceStatusError>

Updates the device status.

fn update_group(
&self,
input: UpdateGroupRequest
) -> RusotoFuture<UpdateGroupResponse, UpdateGroupError>

Updates the specified group with the specified attributes.

Requires developer credentials.

fn update_identity_provider(
&self,
input: UpdateIdentityProviderRequest
) -> RusotoFuture<UpdateIdentityProviderResponse, UpdateIdentityProviderError>

Updates identity provider information for a user pool.

fn update_resource_server(
&self,
input: UpdateResourceServerRequest
) -> RusotoFuture<UpdateResourceServerResponse, UpdateResourceServerError>

Updates the name and scopes of resource server. All other fields are read-only.

fn update_user_attributes(
&self,
input: UpdateUserAttributesRequest
) -> RusotoFuture<UpdateUserAttributesResponse, UpdateUserAttributesError>

Allows a user to update a specific attribute (one at a time).

fn update_user_pool(
&self,
input: UpdateUserPoolRequest
) -> RusotoFuture<UpdateUserPoolResponse, UpdateUserPoolError>

Updates the specified user pool with the specified attributes.

fn update_user_pool_client(
&self,
input: UpdateUserPoolClientRequest
) -> RusotoFuture<UpdateUserPoolClientResponse, UpdateUserPoolClientError>

Allows the developer to update the specified user pool client and password policy.

fn verify_software_token(
&self,
input: VerifySoftwareTokenRequest
) -> RusotoFuture<VerifySoftwareTokenResponse, VerifySoftwareTokenError>

Use this API to register a user's entered TOTP code and mark the user's software token MFA status as "verified" if successful. The request takes an access token or a session string, but not both.

fn verify_user_attribute(
&self,
input: VerifyUserAttributeRequest
) -> RusotoFuture<VerifyUserAttributeResponse, VerifyUserAttributeError>

Verifies the specified user attributes in the user pool.